+++ extensions/sips/src/flip-horizontally.ts
@@ -9,9 +9,9 @@
  */
 
 import flip from "./operations/flipOperation";
+import runOperation from "./operations/runOperation";
 import { Direction } from "./utilities/enums";
 import { getSelectedImages } from "./utilities/utils";
-import runOperation from "./operations/runOperation";
 
 export default async function Command() {
   const selectedImages = await getSelectedImages();