+++ extensions/trakt-manager/src/components/episode-grid.tsx
@@ -1,30 +1,37 @@
-import { Grid, Toast, showToast } from "@raycast/api";
-import { useCachedPromise } from "@raycast/utils";
+import { Action, ActionPanel, Grid, Icon, Keyboard, Toast, showToast } from "@raycast/api";
+import { getFavicon, useCachedPromise } from "@raycast/utils";
 import { setMaxListeners } from "node:events";
-import { useCallback, useEffect, useRef, useState } from "react";
-import { getEpisodes } from "../api/shows";
-import { useEpisodeMutations } from "../hooks/useEpisodeMutations";
-import { APP_MAX_LISTENERS } from "../lib/constants";
-import { EpisodeGridItem } from "./episode-grid-item";
+import { useCallback, useRef, useState } from "react";
+import { initTraktClient } from "../lib/client";
+import { APP_MAX_LISTENERS, IMDB_APP_URL, TRAKT_APP_URL } from "../lib/constants";
+import { getIMDbUrl, getScreenshotUrl, getTraktUrl } from "../lib/helper";
+import { TraktEpisodeListItem } from "../lib/schema";
+import { GenericGrid } from "./generic-grid";
 
-export const EpisodeGrid = ({
-  showId,
-  tmdbId,
-  seasonNumber,
-  slug,
-}: {
-  showId: number;
-  tmdbId: number;
-  seasonNumber: number;
-  slug: string;
-}) => {
+export const EpisodeGrid = ({ showId, seasonNumber, slug }: { showId: number; seasonNumber: number; slug: string }) => {
   const abortable = useRef<AbortController>();
+  const traktClient = initTraktClient();
   const [actionLoading, setActionLoading] = useState(false);
   const { isLoading, data: episodes } = useCachedPromise(
     async (showId: number, seasonNumber: number) => {
       abortable.current = new AbortController();
       setMaxListeners(APP_MAX_LISTENERS, abortable.current?.signal);
-      return await getEpisodes(showId, seasonNumber, abortable.current?.signal);
+
+      const response = await traktClient.shows.getEpisodes({
+        query: {
+          extended: "full,cloud9",
+        },
+        params: {
+          showid: showId,
+          seasonNumber: seasonNumber,
+        },
+        fetchOptions: {
+          signal: abortable.current.signal,
+        },
+      });
+
+      if (response.status !== 200) return [];
+      return response.body;
     },
     [showId, seasonNumber],
     {
@@ -39,13 +46,55 @@ export const EpisodeGrid = ({
       },
     },
   );
-  const { checkInEpisodeMutation, error, success } = useEpisodeMutations(abortable);
+
+  const addEpisodeToHistory = useCallback(async (episode: TraktEpisodeListItem) => {
+    await traktClient.shows.addEpisodeToHistory({
+      body: {
+        episodes: [
+          {
+            ids: {
+              trakt: episode.ids.trakt,
+            },
+            watched_at: new Date().toISOString(),
+          },
+        ],
+      },
+    });
+  }, []);
+
+  const checkInEpisode = useCallback(async (episode: TraktEpisodeListItem) => {
+    await traktClient.shows.checkInEpisode({
+      body: {
+        episodes: [
+          {
+            ids: {
+              trakt: episode.ids.trakt,
+            },
+            watched_at: new Date().toISOString(),
+          },
+        ],
+      },
+    });
+  }, []);
 
   const handleAction = useCallback(
-    async (episode: TraktEpisodeListItem, action: (episode: TraktEpisodeListItem) => Promise<void>) => {
+    async (
+      episode: TraktEpisodeListItem,
+      action: (episode: TraktEpisodeListItem) => Promise<void>,
+      message: string,
+    ) => {
       setActionLoading(true);
       try {
         await action(episode);
+        showToast({
+          title: message,
+          style: Toast.Style.Success,
+        });
+      } catch (e) {
+        showToast({
+          title: (e as Error).message,
+          style: Toast.Style.Failure,
+        });
       } finally {
         setActionLoading(false);
       }
@@ -53,43 +102,48 @@ export const EpisodeGrid = ({
     [],
   );
 
-  useEffect(() => {
-    if (success) {
-      showToast({
-        title: success,
-        style: Toast.Style.Success,
-      });
-    }
-  }, [success]);
-
-  useEffect(() => {
-    if (error) {
-      showToast({
-        title: error.message,
-        style: Toast.Style.Failure,
-      });
-    }
-  }, [error]);
-
   return (
-    <Grid
+    <GenericGrid
       isLoading={isLoading || actionLoading}
-      columns={3}
       aspectRatio="16/9"
       fit={Grid.Fit.Fill}
+      columns={3}
       searchBarPlaceholder="Search for episodes"
-    >
-      {episodes &&
-        episodes.map((episode) => (
-          <EpisodeGridItem
-            key={episode.ids.trakt}
-            episode={episode}
-            tmdbId={tmdbId}
-            seasonNumber={seasonNumber}
-            slug={slug}
-            checkInEpisodeMutation={() => handleAction(episode, checkInEpisodeMutation)}
-          />
-        ))}
-    </Grid>
+      items={episodes}
+      title={(item) => item.title}
+      subtitle={(item) => `Episode ${item.number}`}
+      poster={(item) => getScreenshotUrl(item.images, "episode.png")}
+      keyFn={(item, index) => `${item.ids.trakt}-${index}`}
+      actions={(item) => (
+        <ActionPanel>
+          <ActionPanel.Section>
+            <Action.OpenInBrowser
+              icon={getFavicon(TRAKT_APP_URL)}
+              title="Open in Trakt"
+              url={getTraktUrl("episode", slug, seasonNumber, item.number)}
+            />
+            <Action.OpenInBrowser
+              icon={getFavicon(IMDB_APP_URL)}
+              title="Open in IMDb"
+              url={getIMDbUrl(item.ids.imdb)}
+            />
+          </ActionPanel.Section>
+          <ActionPanel.Section>
+            <Action
+              title="Check-in"
+              icon={Icon.Checkmark}
+              shortcut={Keyboard.Shortcut.Common.ToggleQuickLook}
+              onAction={() => handleAction(item, checkInEpisode, "Episode checked-in")}
+            />
+            <Action
+              title="Add to History"
+              icon={Icon.Clock}
+              shortcut={Keyboard.Shortcut.Common.Duplicate}
+              onAction={() => handleAction(item, addEpisodeToHistory, "Episode added to history")}
+            />
+          </ActionPanel.Section>
+        </ActionPanel>
+      )}
+    />
   );
 };